[FLINK-29655][hotfix] Move ResourceLifecycleState to api module
