+++ flink-kubernetes-operator/src/main/java/org/apache/flink/kubernetes/operator/listener/AuditUtils.java
@@ -19,7 +19,6 @@ package org.apache.flink.kubernetes.operator.listener;
 
 import org.apache.flink.kubernetes.operator.api.AbstractFlinkResource;
 import org.apache.flink.kubernetes.operator.api.status.CommonStatus;
-import org.apache.flink.kubernetes.operator.metrics.OperatorMetricUtils;
 
 import io.fabric8.kubernetes.api.model.Event;
 import lombok.NonNull;
@@ -42,7 +41,7 @@ public class AuditUtils {
     }
 
     private static String format(@NonNull CommonStatus<?> status) {
-        var lifeCycleState = OperatorMetricUtils.getLifecycleState(status);
+        var lifeCycleState = status.getLifecycleState();
         return String.format(
                 ">>> Status | %-7s | %-15s | %s ",
                 StringUtils.isEmpty(status.getError()) ? "Info" : "Error",
